rename "plone" package to "ploneintegration" for awoiding import conflicts.
