--- errbit/plone/__init__.py
@@ -1,5 +1,0 @@
-from errbit.plone.loghandler import ErrbitLoggingHandler
-import logging
-
-
-logging.root.addHandler(ErrbitLoggingHandler())
