+++ errbit/ploneintegration/__init__.py
@@ -1,0 +1,5 @@
+from errbit.ploneintegration.loghandler import ErrbitLoggingHandler
+import logging
+
+
+logging.root.addHandler(ErrbitLoggingHandler())
